# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FaroBot Studio — AI-powered code correction assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  model: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title: string;
  model: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  fileContext?: string | null;
  createdAt: string;
}

export interface ConversationDetail {
  id: number;
  title: string;
  model: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
  /** @nullable */
  fileContext?: string | null;
  /** @nullable */
  githubFile?: string | null;
}

export interface AiModel {
  id: string;
  name: string;
  description: string;
  category: string;
}

export interface GithubStatus {
  connected: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface GithubRepo {
  id: number;
  name: string;
  fullName: string;
  /** @nullable */
  description?: string | null;
  private: boolean;
  /** @nullable */
  language?: string | null;
  updatedAt: string;
}

export type GithubFileType = typeof GithubFileType[keyof typeof GithubFileType];


export const GithubFileType = {
  file: 'file',
  dir: 'dir',
} as const;

export interface GithubFile {
  name: string;
  path: string;
  type: GithubFileType;
  /** @nullable */
  size?: number | null;
  sha: string;
}

export interface FileContent {
  path: string;
  content: string;
  sha: string;
  encoding: string;
}

export interface FileRequest {
  owner: string;
  repo: string;
  path: string;
}

export interface DirRequest {
  owner: string;
  repo: string;
  path: string;
}

export interface FileUpdate {
  owner: string;
  repo: string;
  path: string;
  content: string;
  sha: string;
  message: string;
}

export interface Settings {
  defaultModel: string;
  /** @nullable */
  huggingfaceToken?: string | null;
  systemPrompt: string;
}

export interface SettingsUpdate {
  defaultModel?: string;
  /** @nullable */
  huggingfaceToken?: string | null;
  systemPrompt?: string;
}