# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FaroBot Studio — AI-powered code correction assistant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /github/repos/{owner}/{repo}/contents:
    get:
      operationId: listRepoContents
      tags: [github]
      summary: List root-level contents of a repository
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repository contents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GithubFile"
  /github/file:
    post:
      operationId: fetchFileContent
      tags: [github]
      summary: Fetch file content from GitHub (POST to avoid query-param collisions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRequest"
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
  /github/file/update:
    post:
      operationId: updateFileContent
      tags: [github]
      summary: Update file content on GitHub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
  /github/dir:
    post:
      operationId: listDirContents
      tags: [github]
      summary: List contents of a directory in a repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DirRequest"
      responses:
        "200":
          description: Directory contents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GithubFile"
...
components:
  schemas:
...
    GithubFile:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, dir]
        size:
          type: ["integer", "null"]
        sha:
          type: string
      required:
        - name
        - path
        - type
        - sha
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        sha:
          type: string
        encoding:
          type: string
      required:
        - path
        - content
        - sha
        - encoding
    FileRequest:
      type: object
      properties:
        owner:
          type: string
        repo:
          type: string
        path:
          type: string
      required:
        - owner
        - repo
        - path
    DirRequest:
      type: object
      properties:
        owner:
          type: string
        repo:
          type: string
        path:
          type: string
      required:
        - owner
        - repo
        - path
    FileUpdate:
      type: object
      properties:
        owner:
          type: string
        repo:
          type: string
        path:
          type: string
        content:
          type: string
        sha:
          type: string
        message:
          type: string
      required:
        - owner
        - repo
        - path
        - content
        - sha
        - message
...