# lib/api-zod/src/generated/types/fileContent.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FaroBot Studio — AI-powered code correction assistant
 * OpenAPI spec version: 0.1.0
 */

export interface FileContent {
  path: string;
  content: string;
  sha: string;
  encoding: string;
}