# lib/api-zod/src/generated/types/fileRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FaroBot Studio — AI-powered code correction assistant
 * OpenAPI spec version: 0.1.0
 */

export interface FileRequest {
  owner: string;
  repo: string;
  path: string;
}